refactor(SectioningContent): useSectioningWrapperをclient/hooks/から移動する
isStyledComponentの除去によりuseSectioningWrapperはclient専用APIに
依存しなくなったため、client/hooks/に置く理由がなくなった。
SectioningContent直下へ移動し、空になったclient/hooks/を削除した。

hookはモジュールスコープでは何も実行せず、'use client'を持つ
SectioningFragmentを参照して返すだけのため、呼び出し元がclientグラフ
内であれば問題ない（CLAUDE.mdの原則どおり）。

Panel等から実際に'use client'を外せるかは別途実Next.jsビルドでの
検証が必要（このコミットのスコープ外）。
