--- packages/smarthr-ui/src/components/SectioningContent/client/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useSectionWrapper } from './useSectioningWrapper'
