+++ packages/smarthr-ui/src/components/SectioningContent/index.ts
@@ -1,3 +1,3 @@
 export { Article, Aside, Nav, Section } from './SectioningContent'
 export { LevelContext } from './client/components'
-export { useSectionWrapper } from './client/hooks'
+export { useSectionWrapper } from './useSectioningWrapper'
